dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
websockets==12.0
httpx==0.25.1
PyJWT==2.8.0